% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDummies.R
\docType{data}
\name{dummies_example}
\alias{dummies_example}
\title{United States Census data for 2015}
\format{A data frame with 100,000 rows and 17 variables:
\describe{
  \item{YEAR}{Year of the survey}
  \item{STATEFIP}{FIPS ID for the state}
  \item{COUNTYFIPS}{FIPS ID for the county}
  \item{OWNERSHP}{Respondent's ownership status for their home}
  \item{PERWT}{Survey weight for the respondent}
  \item{NCHILD}{Number of children in the home}
  \item{SEX}{Sex of the respondent}
  \item{AGE}{Age of the respondent}
  \item{MARST}{Marital status of the respondent}
  \item{RACE}{Race of the respondent}
  \item{HISPAN}{If the respondent is hispanic}
  \item{BPL}{Birthplace of the respondent}
  \item{EDUC}{Education of the respondent}
  \item{EMPSTAT}{Employment status of the respondent}
  \item{FTOTINC}{Family yearly income}
  \item{POVERTY}{If they are below the poverty line}
  \item{MOVEDIN}{When the respondent moved into their home}
}}
\source{
\url{https://usa.ipums.org/usa-action/variables/group}
}
\usage{
dummies_example
}
\description{
A dataset containing Census results from the American
Community Survey 2015
}
\keyword{datasets}
