% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDummies.R
\docType{data}
\name{dummy_rows_example}
\alias{dummy_rows_example}
\title{National Incident-Based Reporting System crime data}
\format{A data frame with 4,139 rows and 6 variables:
\describe{
  \item{state}{State}
  \item{year}{Year}
  \item{sexofoffender}{Sex of the offender}
  \item{offender_age}{Age of the offender}
  \item{raceofoffender}{Race of the offender}
  \item{total_assault}{Number of assault crimes}
}}
\source{
\url{http://www.icpsr.umich.edu/icpsrweb/NACJD/studies/3449}
}
\usage{
dummy_rows_example
}
\description{
A dataset crime information from the 2000 NIBRS
}
\keyword{datasets}
