% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_cols.R
\name{dummy_cols}
\alias{dummy_cols}
\title{Fast creation of dummy variables}
\usage{
dummy_cols(data, select_columns = NULL, remove_first_dummy = FALSE)
}
\arguments{
\item{data}{An object with the data set you want to make dummy columns from.}

\item{select_columns}{Vector of column names that you want to create dummy variables from.
If NULL (default), uses all character and factor columns.}

\item{remove_first_dummy}{Removes the first dummy of every variable that only n-1 Dummies remain.
This avoids multicollinearity issues in models.}
}
\value{
A data.frame with same number of rows as inputted data and original
columns plus the newly created dummy columns.
}
\description{
Fast creation of dummy variables
}
